pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opsforge"
version = "1.0.0"
description = "A forge of essential tools for DevOps engineers and system administrators"
readme = "README.md"
authors = [
    {name = "Thomas Vincent", email = "info@thomasvincent.xyz"}
]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
keywords = ["devops", "sysadmin", "monitoring", "automation", "dns", "http", "filesystem"]
dependencies = [
    "requests>=2.28.2",
    "click>=8.1.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "dnspython>=2.3.0",
    "paramiko>=3.0.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.2.0",
    "pylint>=2.17.0",
    "pre-commit>=3.3.1",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.2.0",
]
gsheets = [
    "gspread>=5.9.0",
    "oauth2client>=4.1.3",
]
zenoss = [
    "zeep>=4.2.1",
]

[project.urls]
"Homepage" = "https://github.com/thomasvincent/opsforge"
"Bug Tracker" = "https://github.com/thomasvincent/opsforge/issues"
"Source Code" = "https://github.com/thomasvincent/opsforge"
"Documentation" = "https://opsforge.readthedocs.io"

[project.scripts]
opsforge = "opsforge.cli:main"
opsforge-http500 = "opsforge.http.http500:main"
opsforge-readonly = "opsforge.filesystem.readonly:main"
opsforge-dns = "opsforge.dns.manager:main"

[tool.black]
# Google Python Style Guide compatible black configuration
line-length = 80
target-version = ['py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # Common directories to exclude
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | venv
  | env
)/
'''

[tool.mypy]
# Google Python Style Guide compatible mypy configuration
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

# Per-module options
[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.pytest.ini_options]
# Pytest configuration
minversion = "6.0"
addopts = "-ra -q --strict-markers --cov=opsforge --cov=utility_scripts --cov-report=term-missing"
testpaths = [
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
branch = true
source = ["opsforge", "utility_scripts"]
omit = ["*/tests/*", "*/migrations/*", "*/__pycache__/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.ruff]
line-length = 80
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]